Renames resource generator controller generator
