+++ lib/generators/godmin/policy/policy_generator.rb
@@ -1,7 +1,7 @@
 class Godmin::PolicyGenerator < Rails::Generators::NamedBase
   source_root File.expand_path("../templates", __FILE__)
 
-  def create_controller_files
+  def create_policy
     template "policy.rb", File.join("app/policies", class_path, "#{file_name}_policy.rb")
   end
 end
